# openapi.yaml
openapi: 3.0.3
info:
  title: FastAPI PostgreSQL Demo
  description: |
    A comprehensive FastAPI application that provides messaging capabilities between agents,
    command-line interface operations, and user management with PostgreSQL backend.
    
    ## Features
    - Agent management and messaging system
    - Message threading and conversations
    - Command-line interface integration
    - User management
    - Real-time message tracking with read receipts
    
    ## Security
    This API implements various security measures:
    - **Authentication**: API key-based authentication via X-API-Key header
    - **Input validation and sanitization**: Comprehensive Pydantic model validation
    - **SQL injection prevention**: Protected through SQLAlchemy ORM
    - **Command injection prevention**: Secure shell escaping in CLI operations
    - **Data privacy**: Agent access control and message isolation
    - **Information disclosure protection**: Proper error handling and response sanitization
    - **Transport security**: HTTPS required for production use
    
    **Important Security Notice:**
    - This OpenAPI specification does NOT contain any actual API key values
    - All sensitive information in examples has been redacted (marked with *************)
    - API keys must be obtained through secure channels and stored securely
    - Never commit API keys to version control or include in client-side code
  version: 1.0.0
  contact:
    name: API Support
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /users/{user_id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user from the database by their ID
      operationId: getUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: The ID of the user to retrieve
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. 
        
        **Security Notice:** 
        - API keys should never be included in client-side code or version control
        - Use environment variables or secure configuration management
        - Rotate API keys regularly
        - API keys MUST be transmitted over HTTPS only in production
        - Example header: `X-API-Key: ************`
        
        **HTTPS Requirement:**
        - All API requests MUST use HTTPS in production environments
        - HTTP connections are rejected in production to prevent cleartext API key transmission
        - Local development may use HTTP for testing purposes only
        
        **Note:** This specification does not contain actual API key values for security reasons.
  schemas:
    UserResponse:
      type: object
      required:
        - id
        - name
        - email
        - created_at
      properties:
        id:
          type: integer
          description: User ID
          example: 1
        name:
          type: string
          description: User name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User email address
          example: "john.doe@example.com"
        created_at:
          type: string
          format: date-time
          description: User creation timestamp
          example: "2023-01-01T12:00:00Z"
...
    HTTPError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error detail message
          example: "Item not found"
...